+++ src/rust/src/lazy/dsl.rs
@@ -1440,14 +1440,26 @@ impl RPolarsExpr {
         ))
     }
 
-    pub fn dt_convert_time_zone(&self, tz: String) -> Self {
-        self.0.clone().dt().convert_time_zone(tz).into()
+    pub fn dt_convert_time_zone(&self, time_zone: Robj) -> RResult<RPolarsExpr> {
+        Ok(RPolarsExpr(
+            self.0
+                .clone()
+                .dt()
+                .convert_time_zone(robj_to!(String, time_zone)?)
+                .into(),
+        ))
     }
 
-    pub fn dt_replace_time_zone(&self, tz: Nullable<String>, ambiguous: Robj) -> RResult<Self> {
+    pub fn dt_replace_time_zone(
+        &self,
+        time_zone: Nullable<String>,
+        ambiguous: Robj,
+        non_existent: Robj,
+    ) -> RResult<Self> {
         Ok(RPolarsExpr(self.0.clone().dt().replace_time_zone(
-            tz.into_option(),
+            time_zone.into_option(),
             robj_to!(PLExpr, ambiguous)?,
+            robj_to!(NonExistent, non_existent)?,
         )))
     }