+++ NEWS.md
@@ -2,14 +2,53 @@
 
 ## Polars R Package (development version)
 
-### Breaking changes
+### Breaking changes due to Rust-polars update
 
+- rust-polars is updated to 0.38.3 (#937).
+  - New argument `non_existent` in `$dt$replace_time_zone()` to specify what should
+    happen when a datetime doesn't exist.
+  - In rolling aggregation functions (such as `$rolling_mean()`), the default
+    value of argument `closed` now is `NULL`. Using `closed` with a fixed
+    `window_size` now throws an error.
+
+### Other breaking changes
+
+- Several functions have been rewritten to match the behavior of Python Polars.
+  - In `pl$Series()`, the first argument `x` is renamed to `values` (#933).
+  - `pl$implode(...)` is rewritten to be a syntactic sugar for `pl$col(...)$implode()` (#923).
+  - Unify names of input/output function arguments (935).
+    - All arguments except the first argument must be named arguments.
+    - In `pl$read_*` and `pl$scan_*` functions, the first argument is now `source`.
+    - In `<DataFrame>$write_*` functions, the first argument is now `file`.
+    - In `<LazyFrame>$sink_*` functions, the first argument is now `path`.
+  - In `$dt$convert_time_zone()` and `$dt$replace_time_zone()`, the `tz` argument
+    is renamed to `time_zone` (#944).
+  - In `$dt$replace_time_zone()`, all arguments except `time_zone` must be named
+    arguments (#944).
+  - `pl$date_range()` is completely rewritten (#950).
+    - The argument `end` must be specified.
+    - The default value of `interval` is changed to `"1d"`.
+    - The unused argument (not working in recent versions) `explode` is removed.
+    - The arguments `closed`, `time_unit`, and `time_zone` must be named arguments.
+    - No longer accepts numeric values to `start` and `end`.
+    - The usage of `pl$date_range()` to create a range of `Datetime` data type is deprecated.
+      `pl$date_range()` will always create a range of `Date` data type in the future.
+      Please use `pl$datetime_range()` if you want to create a range of `Datetime` instead.
 - The argument `columns` in `$drop()` is removed. `$drop()` now accepts several
   character scalars, such as `$drop("a", "b", "c")` (#912).
 - In `pl$col()`, the `name` argument is removed, and the `...` argument no longer
   accepts a list of characters and `RPolarsSeries` class objects (#923).
-- `pl$implode(...)` is rewritten to be a syntactic sugar for `pl$col(...)$implode()` (#923).
 - Removed `$argsort()` which was an old alias for `$arg_sort()` (#930).
+- Removed `pl$expr_to_r()` which was an alias for `$to_r()` (#938).
+- `<Series>$to_r_list()` is renamed `<Series>$to_list()` (#938).
+- Removed `<Series>$to_r_vector()` which was an old alias for
+  `<Series>$to_vector()` (#938).
+- In `$str$strptime()`, `$str$to_date()`, `$str$to_datetime()`, and
+  `$str$to_time()`, all arguments (except the first one) must be named (#939).
+- In `$str$strptime()`, the argument `datatype` is renamed `dtype` (#939).
+- `$unique()` for `DataFrame` and `LazyFrame` have several changes (#953):
+  - New default value `"any"` for argument `keep`.
+  - Arguments `keep` and `maintain_order` must be named.
 
 ### New features
 
@@ -17,6 +56,18 @@
   Expr of class datetime, date, and time via columns and literals (#918).
 - New function `pl$arg_where()` to get the indices that match a condition (#922).
 - New function `is_polars_dtype()` (#927).
+- New function `pl$datetime_range()` (#950).
+- New method `<LazyFrame>$to_dot()` to print the query plan of a LazyFrame with
+  graphviz dot syntax (#928).
+- Argument `ambiguous` can now take the value `"null"` to convert ambigous
+  datetimes to null values (#937).
+- New function `pl$mean_horizontal()` (#959).
+
+### Bug fixes
+
+- The `join_nulls` and `validate` arguments of `<DataFrame>$join()` now work
+  correctly (#945).
+- Export the `Duration` datatype (#955).
 
 ## Polars R Package 0.15.1